requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
Deprecated==1.3.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
feedparser==6.0.12
greenlet==3.2.4
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
limits==4.2
packaging==24.2
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pypdf==6.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sgmllib3k==1.0.0
slowapi==0.1.9
sniffio==1.3.1
jinja2==3.1.5
soupsieve==2.8.3
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.1
